+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for Multi-entry API, [PR-60](https://github.com/reductstore/reduct-rs/pull/60)
 - Add multi-entry query link support (API v1.18+), [PR-61](https://github.com/reductstore/reduct-rs/pull/61)
 
+### Removed
+
+- each_n`, `each_s` and `limit` parameters in `QueryBuilder` and `ReplicationSettings` are deprecated, [PR-62](https://github.com/reductstore/reduct-rs/pull/62)
+
 ## 1.17.2 - 2025-12-17
 
 ### Fixed